# src/main/java/com/kshrd/assessment/docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Assessment Management API",
    "description": "API for managing assessments, exams, sections, questions, and student assessments",
    "version": "v4"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/v4/sections/{sectionId}": {
      "get": {
        "tags": [
          "Section Management"
        ],
        "summary": "Get section by ID",
        "description": "Retrieves a specific section by its unique identifier including section name and associated assessment ID",
        "operationId": "getSection",
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseSectionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Section Management"
        ],
        "summary": "Update section",
        "description": "Updates the name of an existing section within an exam",
        "operationId": "updateSection",
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SectionUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Section Management"
        ],
        "summary": "Delete section",
        "description": "Deletes a section and all its associated questions from an exam",
        "operationId": "deleteSection",
        "parameters": [
          {
            "name": "sectionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v4/questions/{questionId}": {
      "get": {
        "tags": [
          "Question Management"
        ],
        "summary": "Get question by ID",
        "description": "Retrieves a specific question by its unique identifier including question type, content, image, and associated section ID",
        "operationId": "getQuestion",
        "parameters": [
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseQuestionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Question Management"
        ],
        "summary": "Update question",
        "description": "Updates the details of an existing question including type, image, and question content",
        "operationId": "updateQuestion",
        "parameters": [
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Question Management"
        ],
        "summary": "Delete question",
        "description": "Deletes a question from a section",
        "operationId": "deleteQuestion",
        "parameters": [
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v4/answers/question/{questionId}": {
      "get": {
        "tags": [
          "Answer Management"
        ],
        "summary": "Get answer for question",
        "description": "Retrieves the answer for a specific question",
        "operationId": "getAnswerByQuestion",
        "parameters": [
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseAnswerResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "SectionUpdateRequest": {
        "type": "object",
        "properties": {
          "sectionName": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "sectionName"
        ]
      },
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {

          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "errorCode": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
      "QuestionUpdateRequest": {
        "type": "object",
        "properties": {
          "questionType": {
            "type": "string",
            "enum": [
              "Multiple Choice",
              "True/False",
              "Long Answer",
              "Code Problem"
            ]
          },
          "image": {
            "type": "string"
          },
          "questionContent": {
            "type": "object",
            "additionalProperties": {

            }
          },
          "points": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "questionContent",
          "questionType"
        ]
      },
...
      "QuestionResponse": {
        "type": "object",
        "properties": {
          "questionId": {
            "type": "string",
            "format": "uuid"
          },
          "questionType": {
            "type": "string",
            "enum": [
              "Multiple Choice",
              "True/False",
              "Long Answer",
              "Code Problem"
            ]
          },
          "image": {
            "type": "string"
          },
          "questionContent": {
            "type": "object",
            "additionalProperties": {

            }
          },
          "sectionId": {
            "type": "string",
            "format": "uuid"
          },
          "points": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "SectionResponse": {
        "type": "object",
        "properties": {
          "sectionId": {
            "type": "string",
            "format": "uuid"
          },
          "sectionName": {
            "type": "string"
          },
          "assessmentId": {
            "type": "string",
            "format": "uuid"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionResponse"
            }
          }
        }
      },
...
      "AnswerResponse": {
        "type": "object",
        "properties": {
          "answerId": {
            "type": "string",
            "format": "uuid"
          },
          "studentId": {
            "type": "string",
            "format": "uuid"
          },
          "questionId": {
            "type": "string",
            "format": "uuid"
          },
          "assessmentId": {
            "type": "string",
            "format": "uuid"
          },
          "answer": {
            "type": "object",
            "additionalProperties": {

            }
          },
          "score": {
            "type": "number",
            "format": "double"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ApiResponseAnswerResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/AnswerResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "errorCode": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
...
      "ApiResponseSectionResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/SectionResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "errorCode": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
      "ApiResponseQuestionResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/QuestionResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "errorCode": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}